requirements.txt
Flask==3.1.0
spotipy==2.25.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.1.0